feat(gridfs): implement GridfsCore for GridFS service
